+++ CHANGELOG.md
@@ -8,6 +8,17 @@
 
 ### Fixed
 
+## 1.3.3 - [2023-06-13]
+
+### Fixed
+
+- Paradata : fix the method for calculating sessions time
+- Exclude boolean variables from length errors
+
+## 1.3.2 - [2023-05-30]
+
+### Fixed
+- Security issue by updating Springboot
 
 ## 1.3.1 - [2023-05-02]